Fixed header colour related tests
